+++ ldap4j-samples/src/main/java/hu/gds/ldap4j/samples/NettyLdapCodecSample.java
@@ -21,8 +21,9 @@ import io.netty.bootstrap.Bootstrap;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.ChannelOption;
 import io.netty.channel.EventLoopGroup;
+import io.netty.channel.MultiThreadIoEventLoopGroup;
 import io.netty.channel.SimpleChannelInboundHandler;
-import io.netty.channel.nio.NioEventLoopGroup;
+import io.netty.channel.nio.NioIoHandler;
 import io.netty.channel.socket.DuplexChannel;
 import io.netty.channel.socket.nio.NioSocketChannel;
 import java.net.InetSocketAddress;
@@ -123,7 +124,7 @@ public class NettyLdapCodecSample {
     public static void main(String[] args) throws Throwable {
         System.out.println("netty codec sample");
         // new thread pool
-        EventLoopGroup eventLoopGroup=new NioEventLoopGroup(8);
+        EventLoopGroup eventLoopGroup=new MultiThreadIoEventLoopGroup(8, NioIoHandler.newFactory());
         try {
             ChannelHandler channelHandler=new ChannelHandler();
             DuplexChannel channel=(DuplexChannel)new Bootstrap()